stitcher/requirements.txt
# Note: python-tk (tkinter) usually comes with Python by default
# inih and exiv2 are system libraries, may need separate installation

piexif

# Computer Vision and Image Processing
opencv-python>=4.5.0
numpy>=1.19.0
imageio>=2.9.0
rawpy>=0.16.0
pyexiv2>=2.8.0

# SVG Processing (requires GTK3 runtime on Windows)
cairosvg>=2.5.2
cairocffi>=1.2.0
tinycss2>=1.1.0
cssselect2>=0.4.0
defusedxml>=0.7.0
webencodings>=0.5.0

# Background Removal (rembg functionality)
rembg>=2.0.0
torch>=1.9.0
torchvision>=0.10.0

# Machine Learning Models (for rembg)
onnxruntime>=1.8.0

# Image Processing Libraries
Pillow>=8.0.0
scikit-image>=0.18.0

# Data Processing
scipy>=1.6.0

# Utilities
tqdm>=4.60.0
requests>=2.25.0

# Excel export for measurement comparisons
pandas>=1.3.0
openpyxl>=3.0.0